requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
blinker==1.7.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.4.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==4.2.6
django-bootstrap-v5==1.0.11
django-browser-reload==1.12.0
django-cors-headers==4.3.0
django-markdown==0.8.4
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
django-tailwind==3.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Flask==3.0.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mistune==3.0.2
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rich==13.6.0
rpds-py==0.13.1
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tailwind==3.1.5b0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1
whitenoise==6.6.0
